+++ activerecord/test/schema/schema.rb
@@ -36,6 +36,7 @@ ActiveRecord::Schema.define do
     t.string :preferences, null: true, default: "", limit: 1024
     t.string :json_data, null: true, limit: 1024
     t.string :json_data_empty, null: true, default: "", limit: 1024
+    t.json :json_native_data
     t.text :params
     t.references :account
     t.json :json_options